# api/gateway/analysis/api.yml
swagger: '2.0'
info:
  version: '1.0.0'  # API version, not the SAR code version
  title: panther-analysis-api
  description: Panther Python policies/rules
  contact:
    name: Panther Labs
    url: https://runpanther.io/about
    email: support@runpanther.io
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json

# The name of the CloudFormation resource for the Lambda handler function
...
parameters:  # Parameters shared across multiple methods
  policyId:
    name: policyId
    in: query
    description: Unique ASCII policy identifier
    required: true
    type: string
    pattern: '[a-zA-Z0-9\-\.: ]{1,200}'
  ruleId:
    name: ruleId
    in: query
    description: Unique ASCII rule identifier
    required: true
    type: string
    pattern: '[a-zA-Z0-9\-\. ]{1,200}'
...
paths:
  /policy:
    # The UI policy detail shows all details for an individual policy.
    #
    # Example: GET /policy ? policyId=BucketEncryptionEnabled
    #
    # Response: {
    #     "autoRemediationId":         "",
    #     "autoRemediationParameters": {},
    #     "body":                      "def policy(resource): return True",
    #     "complianceStatus":          "PASS",
    #     "createdAt":                 "2019-08-26T00:00:00.000Z",
    #     "createdBy":                 "5f54cf4a-ec56-44c2-83bc-8b742600f307",
    #     "description":               "This policy is always compliant",
    #     "displayName":               "My Test Policy",
    #     "enabled":                   true,
    #     "id":                        "BucketEncryptionEnabled",
    #     "lastModified":              "2019-08-26T00:00:00.000Z",
    #     "lastModifiedBy":            "5f54cf4a-ec56-44c2-83bc-8b742600f307",
    #     "reference":                 "https://example.com",
    #     "resourceTypes":             ["AWS.S3.Bucket"],
    #     "runbook":                   "No need to do anything",
    #     "severity":                  "INFO",
    #     "suppressions":              ["arn:aws:s3:::panther-public-*"],
    #     "tags":                      [],
    #     "tests": [
    #         {
    #             "expectedResult": true,
    #             "name": "TestTheFirst",
    #             "resource": {
    #                 "hello": "world"
    #             },
    #             "type": "AWS.S3.Bucket"
    #         }
    #     ],
    #     "versionId": "TsKejJ6GGi_KdH65g2iu9bcww8JxkkwI"
    # }
    get:
      operationId: GetPolicy
      summary: Get policy details
      parameters:
        - $ref: '#/parameters/policyId'
        -
          name: versionId
          in: query
          description: Optional version ID to retrieve (for older versions)
          type: string
          pattern: '[a-zA-Z\._0-9]{32}'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Policy'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Policy does not exist
        500:
          description: Internal server error

    # When users create a new policy, the frontend uses this endpoint.
    #
    # This is almost identical to updating a policy, except the policyId must not already exist.
    #
    # Example: (see ModifyPolicy)
    post:
      operationId: CreatePolicy
      summary: Create a new compliance policy
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdatePolicy'
      responses:
        201:
          description: Policy created successfully
          schema:
            $ref: '#/definitions/Policy'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Rule or policy with the given ID already exists
        500:
          description: Internal server error
  /rule:
    # Same as GetPolicy, but for a log analysis rule (which has slightly different parameters).
    get:
      operationId: GetRule
      summary: Get rule details
      parameters:
        - $ref: '#/parameters/ruleId'
        -
          name: versionId
          in: query
          description: Optional version ID to retrieve (for older versions)
          type: string
          pattern: '[a-zA-Z\._0-9]{32}'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Rule'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Rule does not exist
        500:
          description: Internal server error

    # Same as CreatePolicy, but for a log analysis rule.
    post:
      operationId: CreateRule
      summary: Create a new log analysis rule
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateRule'
      responses:
        201:
          description: Rule created successfully
          schema:
            $ref: '#/definitions/Rule'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Rule or policy with the given ID already exists
        500:
          description: Internal server error
...
  /update:
    # When users edit a policy, the frontend sends the changes to this endpoint.
    #
    # This is almost identical to creating a policy, except the policyId must already exist.
    # NOTE: we can't use PATCH because of a limitation in AppSync.
    #
    # Example: POST /update
    # {
    #     "body":     "def policy(resource): return False",
    #     "enabled":  false,
    #     "id":       "MyVeryExcellentPolicy",
    #     "severity": "MEDIUM",
    #     "userId":   "5f54cf4a-ec56-44c2-83bc-8b742600f307"
    # }
    #
    # Response: (see GetPolicy)
    post:
      operationId: ModifyPolicy
      summary: Modify an existing policy
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdatePolicy'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Policy'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Policy not found
        500:
          description: Internal server error
...
definitions:
  Error:
    type: object
    properties:
      message:
        description: Error message
        type: string
    required:
      - message

  ##### Create/Update/Modify #####
  Policy:
    type: object
    properties:
      autoRemediationId:
        $ref: '#/definitions/autoRemediationId'
      autoRemediationParameters:
        $ref: '#/definitions/autoRemediationParameters'
      body:
        $ref: '#/definitions/body'
      complianceStatus:
        $ref: '#/definitions/complianceStatus'
      createdAt:
        $ref: '#/definitions/modifyTime'
      createdBy:
        $ref: '#/definitions/userId'
      description:
        $ref: '#/definitions/description'
      displayName:
        $ref: '#/definitions/displayName'
      enabled:
        $ref: '#/definitions/enabled'
      id:
        $ref: '#/definitions/id'
      lastModified:
        $ref: '#/definitions/modifyTime'
      lastModifiedBy:
        $ref: '#/definitions/userId'
      reference:
        $ref: '#/definitions/reference'
      resourceTypes:
        $ref: '#/definitions/TypeSet'
      runbook:
        $ref: '#/definitions/runbook'
      severity:
        $ref: '#/definitions/severity'
      suppressions:
        $ref: '#/definitions/suppressions'
      tags:
        $ref: '#/definitions/tags'
      tests:
        $ref: '#/definitions/TestSuite'
      versionId:
        $ref: '#/definitions/versionId'
    required:
      - autoRemediationId
      - autoRemediationParameters
      - body
      - complianceStatus
      - createdAt
      - createdBy
      - description
      - displayName
      - enabled
      - id
      - lastModified
      - lastModifiedBy
      - reference
      - resourceTypes
      - runbook
      - severity
      - suppressions
      - tags
      - tests
      - versionId
  TypeSet:
    description: List of resource/log types to which this policy applies
    type: array
    maxItems: 500
    uniqueItems: true
    items:
      type: string
      minLength: 1
  UnitTest:
    type: object
    properties:
      expectedResult:
        $ref: '#/definitions/testExpectedResult'
      name:
        $ref: '#/definitions/testName'
      resource:
        $ref: '#/definitions/testResource'
      resourceType:
        $ref: '#/definitions/testResourceType'
    required:
      - expectedResult
      - name
      - resource
      - resourceType
  TestSuite:
    type: array
    items:
      $ref: '#/definitions/UnitTest'
  UpdatePolicy:
    type: object
    properties:
      autoRemediationId:
        $ref: '#/definitions/autoRemediationId'
      autoRemediationParameters:
        $ref: '#/definitions/autoRemediationParameters'
      body:
        $ref: '#/definitions/body'
      description:
        $ref: '#/definitions/description'
      displayName:
        $ref: '#/definitions/displayName'
      enabled:
        $ref: '#/definitions/enabled'
      id:
        $ref: '#/definitions/id'
      reference:
        $ref: '#/definitions/reference'
      resourceTypes:
        $ref: '#/definitions/TypeSet'
      runbook:
        $ref: '#/definitions/runbook'
      severity:
        $ref: '#/definitions/severity'
      suppressions:
        $ref: '#/definitions/suppressions'
      tags:
        $ref: '#/definitions/tags'
      tests:
        $ref: '#/definitions/TestSuite'
      userId:
        $ref: '#/definitions/userId'
    required:
      - body
      - enabled
      - id
      - severity
      - userId

  ##### DeletePolicies #####
...
  Rule:
    type: object
    properties:
      body:
        $ref: '#/definitions/body'
      createdAt:
        $ref: '#/definitions/modifyTime'
      createdBy:
        $ref: '#/definitions/userId'
      description:
        $ref: '#/definitions/description'
      displayName:
        $ref: '#/definitions/displayName'
      enabled:
        $ref: '#/definitions/enabled'
      id:
        $ref: '#/definitions/id'
      lastModified:
        $ref: '#/definitions/modifyTime'
      lastModifiedBy:
        $ref: '#/definitions/userId'
      logTypes:
        $ref: '#/definitions/TypeSet'
      reference:
        $ref: '#/definitions/reference'
      runbook:
        $ref: '#/definitions/runbook'
      severity:
        $ref: '#/definitions/severity'
      tags:
        $ref: '#/definitions/tags'
      tests:
        $ref: '#/definitions/TestSuite'
      versionId:
        $ref: '#/definitions/versionId'
    required:
      - body
      - createdAt
      - createdBy
      - description
      - displayName
      - enabled
      - id
      - lastModified
      - lastModifiedBy
      - logTypes
      - reference
      - runbook
      - severity
      - tags
      - tests
      - versionId
  UpdateRule:
    type: object
    properties:
      body:
        $ref: '#/definitions/body'
      description:
        $ref: '#/definitions/description'
      displayName:
        $ref: '#/definitions/displayName'
      enabled:
        $ref: '#/definitions/enabled'
      id:
        $ref: '#/definitions/id'
      logTypes:
        $ref: '#/definitions/TypeSet'
      reference:
        $ref: '#/definitions/reference'
      runbook:
        $ref: '#/definitions/runbook'
      severity:
        $ref: '#/definitions/severity'
      tags:
        $ref: '#/definitions/tags'
      tests:
        $ref: '#/definitions/TestSuite'
      userId:
        $ref: '#/definitions/userId'
    required:
      - body
      - enabled
      - id
      - severity
      - userId

  ##### ListRules #####
...
  autoRemediationId:
    description: When a resource fails the policy, trigger the remediation with this ID
    type: string
    maxLength: 200
  autoRemediationParameters:
    description: Configuration parameters passed to the remediation handler
    type: object
    additionalProperties:
      type: string
    maxProperties: 50
...
  body:
    description: Python policy source code
    type: string
    minLength: 10
    maxLength: 1000000  # ~1 MB
  complianceStatus:
    description: Pass/fail compliance status
    type: string
    enum:
      - ERROR    # Policy raised an exception when evaluating a resource
      - FAIL     # Policy failed on at least one resource
      - PASS     # Policy passed for all applicable resources
  description:
    description: Summary of the policy and its purpose
    type: string
    maxLength: 5000
  displayName:
    description: Optional user-friendly policy name
    type: string
    maxLength: 200
  enabled:
    description: True if the policy is currently being evaluated
    type: boolean
  id:
    description: User-specified unique rule/policy ID
    type: string
    pattern: '[a-zA-Z0-9\-\.: ]{1,200}'
  modifyTime:
    description: Policy modification timestamp
    type: string
    format: date-time
  reference:
    description: External documentation motivating the need for this policy
    type: string
    maxLength: 5000
  runbook:
    description: Internal documenation about what to do when a policy fails
    type: string
  suppressions:
    description: >
      List of resource ID regexes that are excepted from this policy.
      The policy will still be evaluated, but failures will not trigger alerts nor remediations.
    type: array
    maxItems: 500
    uniqueItems: true
    items:
      type: string
      maxLength: 1000
  testExpectedResult:
    description: The expected outcome when running a unit test
    type: boolean
  testName:
    description: The name of a unit test
    type: string
    minLength: 1
  testResourceType:
    description: The resource or log type of an individual unit test
    type: string
    minLength: 1
  testResource:
    description: The resource (or log event) the unit test is running against
    type: string
    minLength: 2
  severity:
    description: Policy severity
    type: string
    enum:
      - INFO
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
  tags:
    description: List of tags to categorize the policy
    type: array
    maxItems: 500
    uniqueItems: true
    items:
      type: string
      maxLength: 200
...
  userId:
    description: Panther user ID that created or modified the policy
    type: string
    pattern: '[a-f0-9\-]{36}'
  versionId:
    description: Policy version ID
    type: string
    pattern: '[a-zA-Z\._0-9]{32}'  # 32-character alphanumeric S3 object version